+++ 2024 Projects/C++ SFML/SFML_Projects/Platformer/HelperFunctions.h
@@ -7,4 +7,4 @@ sf::Vector2f lerp(const sf::Vector2f& a, const sf::Vector2f& b, float t);
 
 sf::Vector2f Normalize(sf::Vector2f& vector, int multiplier);
 
-int sign(int x);
+int sign(float x);
